Update core-foundation and cocoa deps

+++ src/backend/metal/src/internal.rs
@@ -299,7 +299,7 @@ impl ImageClearPipes {
             for (i, &format) in key.color_formats.iter().enumerate() {
                 pipeline
                     .color_attachments()
-                    .object_at(i)
+                    .object_at(i as u64)
                     .unwrap()
                     .set_pixel_format(format);
             }
@@ -330,7 +330,7 @@ impl ImageClearPipes {
         for i in 0 .. 1 {
             let mtl_attribute_desc = vertex_descriptor
                 .attributes()
-                .object_at(i)
+                .object_at(i as u64)
                 .expect("too many vertex attributes");
             mtl_attribute_desc.set_buffer_index(0);
             mtl_attribute_desc.set_offset((i * mem::size_of::<[f32; 4]>()) as _);
@@ -426,7 +426,7 @@ impl ImageBlitPipes {
         for i in 0 .. 2 {
             let mtl_attribute_desc = vertex_descriptor
                 .attributes()
-                .object_at(i)
+                .object_at(i as u64)
                 .expect("too many vertex attributes");
             mtl_attribute_desc.set_buffer_index(0);
             mtl_attribute_desc.set_offset((i * mem::size_of::<[f32; 4]>()) as _);
